+++ src/lib/components/Cart/CartEstimateShipping.tsx
@@ -6,7 +6,6 @@ import ShippingMethods from 'src/lib/components/ShippingMethods';
 import GET_ALL_COUNTRIES from 'src/lib/queries/getAllCountries.graphql';
 import SET_SHIPPING_ADDRESSES_MUTATION from 'src/lib/queries/setShippingAddresses.graphql';
 import SET_SHIPPING_METHOD_MUTATION from 'src/lib/queries/setShippingMethod.graphql';
-import { TCartAddressInput } from 'src/lib/types/graphql/Cart';
 import { useCartEstimate } from 'src/peregrine/lib/talons/adeoweb/Cart/useCartEstimate';
 
 const CartEstimateShipping: React.FunctionComponent = () => {
@@ -26,7 +25,7 @@ const CartEstimateShipping: React.FunctionComponent = () => {
         countries,
         regions,
         isValid
-    } = useCartEstimate<TCartAddressInput>({
+    } = useCartEstimate({
         countriesQuery: GET_ALL_COUNTRIES,
         setShippingAddressesOnCartMutation: SET_SHIPPING_ADDRESSES_MUTATION,
         setShippingMethodOnCartMutation: SET_SHIPPING_METHOD_MUTATION
