--- src/peregrine/lib/talons/adeoweb/Cart/useCart.d.ts
@@ -1,18 +1,0 @@
-import { TCartItem } from 'src/lib/types/graphql/CartItem';
-import { ICartState } from 'src/peregrine/lib/store/reducers/adeoweb/cart';
-
-export type TUseCart = {
-    items: TCartItem[];
-    cartState: ICartState;
-    currencyCode: string;
-    handleBeginEditItem: () => void;
-    handleEndEditItem: () => void;
-    isEditingItem: boolean;
-    isLoading: boolean;
-    isUpdatingItem: boolean;
-    isSubmitting: boolean;
-    numItems: number;
-    subtotal: number;
-};
-
-export function useCart(): TUseCart;
