--- src/peregrine/lib/talons/adeoweb/Cart/useCartBody.d.ts
@@ -1,14 +1,0 @@
-import { TCartItem } from 'src/lib/types/graphql/CartItem';
-
-type TUseCartBodyProps = {
-    beginEditItem: () => void;
-    endEditItem: () => void;
-};
-
-export type TUseCartBody = {
-    editItem: TCartItem;
-    handleBeginEditItem: () => void;
-    handleEndEditItem: () => void;
-};
-
-export function useCartBody(props: TUseCartBodyProps): TUseCartBody;
