+++ src/peregrine/lib/talons/adeoweb/Cart/useCartProduct.tsx
@@ -1,12 +1,42 @@
+import { DocumentNode } from 'graphql';
+
 import { useMutation } from '@apollo/react-hooks';
 import { useCallback, useState } from 'react';
 
 import { useAwaitQuery } from '@magento/peregrine/lib/hooks/useAwaitQuery';
 
+import {
+    TCartItem,
+    TSelectedConfigurableOption,
+    TSelectedCustomizableOption
+} from 'src/lib/types/graphql/CartItem';
 import { useCartContext } from 'src/peregrine/lib/context/adeoweb/cart';
 import { fetchPolicy } from 'src/peregrine/lib/util/adeoweb/fetchPolicy';
+import filterOutNullableValues from 'src/peregrine/lib/util/adeoweb/filterOutNullableValues';
 
-export const useCartProduct = props => {
+type TUseCartProductProps = {
+    createCartMutation: DocumentNode;
+    getCartDetailsQuery: DocumentNode;
+    item: TCartItem;
+    removeItemMutation: DocumentNode;
+};
+
+export type TUseCartProduct = {
+    handleRemoveItem: () => void;
+    isLoading: boolean;
+    image: string;
+    name: string;
+    options: TSelectedConfigurableOption[];
+    customOptions: TSelectedCustomizableOption[];
+    price: number;
+    subtotal: number;
+    urlKey: string;
+    quantity: number;
+};
+
+export const useCartProduct = (
+    props: TUseCartProductProps
+): TUseCartProduct => {
     const {
         createCartMutation,
         getCartDetailsQuery,
@@ -15,16 +45,16 @@ export const useCartProduct = props => {
     } = props;
 
     const {
-        configurable_options: options = [],
-        custom_options: simpleCustomOptions,
-        customizable_options: configCustomOptions,
+        configurable_options,
+        customizable_options,
         product: {
             small_image: smallImage,
             name = '',
             url_key: urlKey = '',
             url_suffix: urlSuffix = ''
         },
-        prices
+        prices,
+        quantity
     } = item;
 
     const price = (prices && prices.price && prices.price.value) || 0;
@@ -54,6 +84,8 @@ export const useCartProduct = props => {
             removeItem
         });
     }, [fetchCartDetails, fetchCartId, item, removeItem, removeItemFromCart]);
+    const options = filterOutNullableValues(configurable_options);
+    const configCustomOptions = filterOutNullableValues(customizable_options);
 
     return {
         handleRemoveItem,
@@ -61,9 +93,10 @@ export const useCartProduct = props => {
         image,
         name,
         options,
-        customOptions: simpleCustomOptions || configCustomOptions || [],
+        customOptions: configCustomOptions || [],
         price,
         subtotal,
-        urlKey: `${urlKey}${urlSuffix}`
+        urlKey: `${urlKey}${urlSuffix}`,
+        quantity
     };
 };
